FIX: Unable to set allowed groups to everyone group (#5)
This update fixes an issue where the copy post button does not show up when it is set to the "everyone" group. This is because the "everyone" group is not a `visible_group`, and only visible groups are attached to the current user's serializer. Hence, we must explicitly check for the everyone group when comparing against the current user's groups.